% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-doc.R
\docType{package}
\name{jds.rmd-package}
\alias{jds.rmd-package}
\title{R Markdown Temlates for Journal of Data Science}
\description{
The \pkg{jds.rmd} package provides customized R Markdown templates and
formats for authoring and submitting articles to Journal of Data Science
(JDS).
}
