% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_article.R
\name{pdf_article}
\alias{pdf_article}
\title{PDF Article for Journal of Data Science}
\usage{
pdf_article(...)
}
\arguments{
\item{...}{FIXME: Other arguments.}
}
\description{
Provides output format to render PDF article for submission to Journal of
Data Science.
}
\examples{
if (interactive() && requireNamespace("rmarkdown")) {
    ## draft from a sample R markdown file
    rmarkdown::draft("jds-sample.Rmd", template = "pdf_article",
                     package = "jds.rmd", edit = FALSE)
    ## produce pdf with the tex source kept for submission
    rmarkdown::render("jds-sample.Rmd")
}
}
