% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_article.R
\name{pdf_article}
\alias{pdf_article}
\title{PDF Article for Journal of Data Science}
\usage{
pdf_article(..., cls = c("jdsart", "jds"))
}
\arguments{
\item{...}{Optional named arguments passed to \code{bookdown::pdf_book()}
and \code{rmarkdown::pdf_document()} other than \code{base_format}.}

\item{cls}{The LaTeX class name. The available choices are \code{jdsart} for
the latest class developed by vtex, or \code{jds} for the deprecated
class.}
}
\value{
The output format that can only be used with
    \code{bookdown::render_book()}.
}
\description{
Provides output format to render PDF article for submission to Journal of
Data Science.
}
\examples{
if (interactive() && requireNamespace("rmarkdown", quietly = TRUE)) {
    ## draft from a sample R markdown file
    jds.rmd::draft("jds-sample.Rmd")
    ## produce pdf with the tex source kept for submission
    rmarkdown::render("jds-sample.Rmd")
}
}
